Port the HTTP router test to the new framework
