+++ libcaf_net/CMakeLists.txt
@@ -56,6 +56,7 @@ caf_add_component(
     caf/net/http/response_header.test.cpp
     caf/net/http/route.cpp
     caf/net/http/router.cpp
+    caf/net/http/router.test.cpp
     caf/net/http/server.cpp
     caf/net/http/server_factory.cpp
     caf/net/http/status.cpp
@@ -115,7 +116,6 @@ caf_add_component(
     net.accept_socket
     net.actor_shell
     net.datagram_socket
-    net.http.router
     net.http.server
     net.ip
     net.length_prefix_framing
